feat: refactor order model to use Prisma enums, add product stock check utility, and enhance product controller responses

+++ src/services/order.service.ts
@@ -1,6 +1,8 @@
 import { Prisma, PrismaClient } from "@prisma/client";
 import { httpStatusCodes } from "../utils/http-status-codes";
-import { OrderStatusEnum } from "../models/order.model";
+import { Order } from "../models/order.model";
+import { ManyOrdersResponse, OneOrderResponse } from "../types";
+import { checkProductsStock } from "../utils/check-products-stock";
 
 const prisma = new PrismaClient();
 
@@ -8,14 +10,23 @@ export class OrderService {
   static async create(
     payload: Prisma.OrderCreateInput,
     products: Prisma.OrderItemCreateManyInput[]
-  ) {
+  ): Promise<OneOrderResponse> {
+    const checkedProducts = await checkProductsStock(products);
+
+    const totalAmount = checkedProducts.reduce((acc, product) => {
+      return acc + Number(product.price) * product.quantity;
+    }, 0);
+
     const newOrder = await prisma.$transaction(async (prisma) => {
       const order = await prisma.order.create({
-        data: payload,
+        data: {
+          ...payload,
+          total: totalAmount,
+        },
       });
 
       await prisma.orderItem.createMany({
-        data: products.map((product) => ({
+        data: checkedProducts.map((product) => ({
           ...product,
           orderId: order.id,
         })),
@@ -24,15 +35,17 @@ export class OrderService {
       return order;
     });
 
-    const data = {
+    const data: Order = {
       ...newOrder,
+      orderItems: checkedProducts,
       total: newOrder.total.toNumber(),
+      status: newOrder.status,
     };
 
-    return { status: httpStatusCodes.CREATED, data };
+    return { status: httpStatusCodes.CREATED, data, error: null };
   }
 
-  static async getById(id: string) {
+  static async getById(id: string): Promise<OneOrderResponse> {
     const order = await prisma.order.findUnique({
       where: {
         id,
@@ -49,19 +62,28 @@ export class OrderService {
     if (!order) {
       return {
         status: httpStatusCodes.NOT_FOUND,
-        data: { message: "Order not found" },
+        data: null,
+        error: "Order not found",
       };
     }
 
-    const data = {
+    const data: Order = {
       ...order,
       total: order.total.toNumber(),
+      orderItems: order.orderItems.map((item) => ({
+        ...item,
+        price: item.price.toNumber(),
+        product: {
+          ...item.product,
+          price: item.product.price.toNumber(),
+        },
+      })),
     };
 
-    return { status: httpStatusCodes.OK, data };
+    return { status: httpStatusCodes.OK, data, error: null };
   }
 
-  static async getByClientId(clientId: string) {
+  static async getByClientId(clientId: string): Promise<ManyOrdersResponse> {
     const orders = await prisma.order.findMany({
       where: {
         clientId,
@@ -78,12 +100,20 @@ export class OrderService {
     const data = orders.map((order) => ({
       ...order,
       total: order.total.toNumber(),
+      orderItems: order.orderItems.map((item) => ({
+        ...item,
+        price: item.price.toNumber(),
+        product: {
+          ...item.product,
+          price: item.product.price.toNumber(),
+        },
+      })),
     }));
 
-    return { status: httpStatusCodes.OK, data };
+    return { status: httpStatusCodes.OK, data, error: null };
   }
 
-  static async getAll() {
+  static async getAll(): Promise<ManyOrdersResponse> {
     const orders = await prisma.order.findMany({
       include: {
         orderItems: {
@@ -97,12 +127,23 @@ export class OrderService {
     const data = orders.map((order) => ({
       ...order,
       total: order.total.toNumber(),
+      orderItems: order.orderItems.map((item) => ({
+        ...item,
+        price: item.price.toNumber(),
+        product: {
+          ...item.product,
+          price: item.product.price.toNumber(),
+        },
+      })),
     }));
 
-    return { status: httpStatusCodes.OK, data };
+    return { status: httpStatusCodes.OK, data, error: null };
   }
 
-  static async update(id: string, payload: Prisma.OrderUpdateInput) {
+  static async update(
+    id: string,
+    payload: Prisma.OrderUpdateInput
+  ): Promise<OneOrderResponse> {
     const order = await prisma.order.findUnique({
       where: {
         id,
@@ -112,7 +153,8 @@ export class OrderService {
     if (!order) {
       return {
         status: httpStatusCodes.NOT_FOUND,
-        data: { message: "Order not found" },
+        data: null,
+        error: "Order not found",
       };
     }
 
@@ -128,21 +170,37 @@ export class OrderService {
       total: updatedOrder.total.toNumber(),
     };
 
-    return { status: httpStatusCodes.OK, data };
+    return { status: httpStatusCodes.OK, data, error: null };
   }
 
   static async delete(id: string) {
-    const order = await prisma.order.delete({
+    const order = await prisma.order.findUnique({
       where: {
         id,
       },
     });
 
+    if (!order) {
+      return {
+        status: httpStatusCodes.NOT_FOUND,
+        data: null,
+        error: "Order not found",
+      };
+    }
+
+    const deletedOrder = await prisma.order.delete({
+      where: {
+        id,
+      },
+      include: {
+        orderItems: true,
+      },
+    });
+
     const data = {
-      ...order,
-      total: order.total.toNumber(),
+      ...deletedOrder,
     };
 
-    return { status: httpStatusCodes.OK, data };
+    return { status: httpStatusCodes.OK, data, error: null };
   }
 }
